requirements.txt
flask==1.1.2
face-recognition==1.3.0
waitress==1.4.4
requests==2.24.0
numpy==1.19.1
pillow==7.2.0
